Rearrange classes
