+++ src/main/java/com/niccholaspage/Fe/API/API.java
@@ -1,7 +1,8 @@
-package com.niccholaspage.Fe;
+package com.niccholaspage.Fe.API;
 
+import com.niccholaspage.Fe.Fe;
+import com.niccholaspage.Fe.Phrase;
 import org.bukkit.ChatColor;
-import com.niccholaspage.Fe.database.Account;
 import java.text.DecimalFormat;
 import java.text.DecimalFormatSymbols;
 import java.util.List;
