+++ src/main/java/com/niccholaspage/Fe/Commands/CleanCommand.java
@@ -1,11 +1,11 @@
-package com.niccholaspage.Fe.command.commands;
+package com.niccholaspage.Fe.Commands;
 
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
 import com.niccholaspage.Fe.Fe;
 import com.niccholaspage.Fe.Phrase;
-import com.niccholaspage.Fe.command.CommandType;
-import com.niccholaspage.Fe.command.SubCommand;
+import com.niccholaspage.Fe.API.CommandType;
+import com.niccholaspage.Fe.API.SubCommand;
 
 public class CleanCommand extends SubCommand {
     private final Fe plugin;
