+++ src/main/java/com/niccholaspage/Fe/Databases/SQLDB.java
@@ -1,8 +1,8 @@
-package com.niccholaspage.Fe.database.databases;
+package com.niccholaspage.Fe.Databases;
 
 import com.niccholaspage.Fe.Fe;
-import com.niccholaspage.Fe.database.Account;
-import com.niccholaspage.Fe.database.Database;
+import com.niccholaspage.Fe.API.Account;
+import com.niccholaspage.Fe.API.Database;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
