+++ src/main/java/com/niccholaspage/Fe/FePlayerListener.java
@@ -1,4 +1,4 @@
-package com.niccholaspage.Fe.listeners;
+package com.niccholaspage.Fe;
 
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
@@ -6,9 +6,8 @@ import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerLoginEvent;
 import org.bukkit.event.player.PlayerQuitEvent;
-import com.niccholaspage.Fe.Fe;
-import com.niccholaspage.Fe.database.Account;
-import com.niccholaspage.Fe.database.Database;
+import com.niccholaspage.Fe.API.Account;
+import com.niccholaspage.Fe.API.Database;
 
 public class FePlayerListener implements Listener
 {
