+++ src/main/java/com/niccholaspage/Fe/VaultHandler.java
@@ -11,7 +11,7 @@ import org.bukkit.event.Listener;
 import org.bukkit.event.server.PluginDisableEvent;
 import org.bukkit.event.server.PluginEnableEvent;
 import org.bukkit.plugin.Plugin;
-import com.niccholaspage.Fe.database.Account;
+import com.niccholaspage.Fe.API.Account;
 import java.util.ArrayList;
 import java.util.List;
